🍇 Begin Sprint Grape ♻️ Change how 'flashAlert' is implemented (refactor this one day to make correct 🤦‍♂️)
