+++ src/index.html
@@ -2,7 +2,7 @@
 <html>
   <head>
     <meta charset="utf-8">
-    <title>Sat - 4/10/21 | Sprint Fig 🥝</title>
+    <title>Sprint Grape 🍇 (Sun 4/25/21)</title>
     <meta name="viewport" content="width=device-width, initial-scale=1">
   </head>
   <body style="visibility: hidden;">
